fix: harden authorization boundaries
