+++ docs/README.md
@@ -1,0 +1,3 @@
+# Docs
+
+Put all design docs in [`designs/`](./designs/). See that folder’s README for layout and naming.
