+++ src/app/(app)/(shell)/admin/events/create-new/page.tsx
@@ -1,22 +1,16 @@
-"use client";
-
-import { useRouter } from "next/navigation";
-
-import { EventCreateModify } from "@/features/admin";
 import { PageContainer } from "@/components/shared/PageContainer";
+import { EventCreateModify } from "@/features/admin";
+import { requireAdmin } from "@/lib/auth/guards";
 
-const AdminCreateEventPage = () => {
-  const router = useRouter();
+export default async function AdminCreateEventPage() {
+  await requireAdmin();
 
   return (
     <PageContainer backHref="/admin/events" backLabel="Back to Events">
       <EventCreateModify
-        onSuccess={() => router.push("/admin/events")}
         title="Create a New Event"
         description="Complete the form below to add a new event to the gallery."
       />
     </PageContainer>
   );
-};
-
-export default AdminCreateEventPage;
+}
