+++ src/features/payments/fulfillment.ts
@@ -6,6 +6,7 @@ import { SquareError, type Payment, type PaymentUpdatedEvent } from "square";
 import type { UserInfoRow } from "@/features/auth";
 import type { CheckoutActionResult, CheckoutRequestInput } from "./types";
 import type { Database, Json } from "@/lib/supabase/database.types";
+import type { MembershipTypeRow } from "@/types/models";
 import {
   fetchEventById,
   fetchEventBySlug,
@@ -78,27 +79,16 @@ function getSquareErrorMessage(
 
 function isMembershipPurchasableForUser(
   user: UserInfoRow,
-  membershipName: string
+  membership: MembershipTypeRow
 ) {
   if (user.membership_type_id || user.membership_pre_ordered_type_id) {
     return false;
   }
 
-  const lowerName = membershipName.toLowerCase();
-
-  if (user.user_type === "ubcStudent" && user.student_number) {
-    return lowerName.includes("explorer") || lowerName.includes("innovator");
-  }
-
-  if (user.user_type === "faculty") {
-    return lowerName.includes("faculty");
-  }
-
-  if (user.user_type === "nonUbc") {
-    return lowerName.includes("non");
-  }
-
-  return false;
+  return (
+    membership.eligible_user_types.includes(user.user_type) &&
+    (user.user_type !== "ubcStudent" || Boolean(user.student_number))
+  );
 }
 
 function getMembershipExpiryIsoString() {
@@ -407,7 +397,7 @@ async function createSquarePaymentForMembership(
     return { error: "Membership plan not found." } as const;
   }
 
-  if (!isMembershipPurchasableForUser(user, membershipType.name)) {
+  if (!isMembershipPurchasableForUser(user, membershipType)) {
     return {
       error: "This membership tier is not available for your account.",
     } as const;
