+++ src/lib/supabase-helpers/users.ts
@@ -2,11 +2,6 @@ import type { DbClient } from "./types";
 import { TABLES } from "./tables";
 import type { UserInfoRow, UserInfoUpdate } from "@/types/models";
 
-interface UserInfoSeed {
-  email?: string;
-  name?: string;
-}
-
 export interface UserInfoContact {
   id: string;
   name: string;
@@ -69,19 +64,6 @@ export async function updateUserInfoById(
   if (error) throw error;
 }
 
-export async function updateUserInfoByEmail(
-  supabase: DbClient,
-  email: string,
-  payload: UserInfoUpdate
-): Promise<void> {
-  const { error } = await supabase
-    .from(TABLES.userInfo)
-    .update(payload)
-    .eq("email", email);
-
-  if (error) throw error;
-}
-
 /**
  * Admin user directory rows: user_info joined with the related membership
  * type name. The `membership_types!membership_type_id(name)` hint pins the
@@ -100,79 +82,4 @@ export async function fetchAdminUserRecords(supabase: DbClient) {
 
   if (error) throw error;
   return data ?? [];
-}
-
-/**
- * Ensures a user_info record exists for the given auth user, creating one if
- * necessary. Returns the user_info auth id used to associate registrations.
- */
-export async function ensureUserInfo(
-  supabase: DbClient,
-  sessionUserId: string,
-  user: UserInfoSeed | null
-): Promise<string> {
-  const { data: existingUserInfo, error: fetchError } = await supabase
-    .from(TABLES.userInfo)
-    .select("id")
-    .eq("auth_user_id", sessionUserId)
-    .maybeSingle();
-
-  if (fetchError) {
-    const errorCode = "code" in fetchError ? fetchError.code : undefined;
-    const errorMessage = fetchError.message || "";
-
-    // PGRST116 is "no rows returned" - expected if user_info doesn't exist yet
-    if (errorCode !== "PGRST116" && !errorMessage.includes("No rows")) {
-      throw new Error(`Failed to load user information: ${errorMessage}`);
-    }
-  }
-
-  if (existingUserInfo?.id) {
-    return existingUserInfo.id;
-  }
-
-  const email = user?.email || "";
-  if (!email) {
-    throw new Error(
-      "Email is required. Please ensure you're logged in with an email address."
-    );
-  }
-
-  const { data: newUserInfo, error: createError } = await supabase
-    .from(TABLES.userInfo)
-    .insert({
-      email: email,
-      name: user?.name || email.split("@")[0] || "User",
-      auth_user_id: sessionUserId,
-      membership_type: "NonUbc",
-      role_access: "basic",
-    })
-    .select("id")
-    .single();
-
-  if (createError) {
-    // Unique constraint violation - record already exists, reuse session id
-    if (createError.code === "23505") {
-      const { data: existingAfterConflict, error: refetchError } = await supabase
-        .from(TABLES.userInfo)
-        .select("id")
-        .eq("auth_user_id", sessionUserId)
-        .maybeSingle();
-
-      if (refetchError) {
-        throw new Error(
-          `Failed to reload your profile after a duplicate record was detected: ${refetchError.message}`
-        );
-      }
-
-      if (existingAfterConflict?.id) {
-        return existingAfterConflict.id;
-      }
-
-      return sessionUserId;
-    }
-    throw new Error(`Failed to set up your profile: ${createError.message}`);
-  }
-
-  return newUserInfo?.id || sessionUserId;
 }
