+++ supabase/rollbacks/20260727020000_atomic_admin_event_delete.sql
@@ -1,0 +1,1 @@
+drop function if exists public.delete_event_atomically(uuid);
