+++ supabase/rollbacks/20260727020500_secure_atomic_admin_event_delete.sql
@@ -1,0 +1,2 @@
+grant execute on function public.delete_event_atomically(uuid)
+  to public, anon, authenticated;
