+++ supabase/rollbacks/20260727021000_remove_browser_admin_user_grants.sql
@@ -1,0 +1,4 @@
+grant update (
+  role_access,
+  membership_type_id
+) on public.user_info to authenticated;
