+++ supabase/rollbacks/20260727031000_add_membership_eligibility.sql
@@ -1,0 +1,2 @@
+alter table public.membership_types
+  drop column if exists eligible_user_types;
